refactor: Replace most Blueprint Dialogs
Replaces most Blueprint Dialogs in the app with a Dialog from shadcn-ui.
Removes the wrapping DarkModeAwareDialog, which is not necessary because
we now apply the dark classes to the body tag.

Still TODO: remove blueprint Dialog from the NewParameterDialog. This
one is a bit trickier because of some z-indexing issues between the
dialog and the popover from ItemClassSelector.